+++ migrate/v0_16/cosmos_test.go
@@ -1,4 +1,4 @@
-package modules
+package v0_16
 
 import (
 	"encoding/json"
